# portal/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: AST Institutional Portal API
  version: 1.0.0
  description: |
    Edge API for the institutional portal (AFC-side admission).

    **Canon alignment (AST-CORE-CANON):**
    - Portal does **not** mint, burn, or hold third-party funds.
    - `processId` = `AST-{INST}-{YYYYMMDD}-{suffix}` (not UUID).
    - Institutional valuation is accepted as given (decimal string); portal does not appraise.
    - Document evidence: package hash + qualified signature attestation (full X.509 chain = follow-on).
    - Economic finality only after Core Orchestrator + PoT on NodeChain.

    **Auth (v1):** session after login (`X-Session-Id`).  
    **Target:** Bearer JWT + mutual TLS (documented; not required in local demo).
servers:
  - url: http://localhost:3100
    description: Local portal edge (BFF)
  - url: https://api.arosfinancialcore.com
    description: Production portal edge (when deployed)
...
security:
  - SessionAuth: []
paths:
...
  /v1/public/processes/{processId}:
    get:
      tags: [public]
      operationId: publicGetProcess
      security: []
      summary: Public process lookup (no registration, no key)
      description: Read-only redacted view. No mint. No NodeChain write.
      parameters:
        - $ref: '#/components/parameters/ProcessId'
      responses:
        '200':
          description: Public process view
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProcessResponse'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /v1/public/nodechain/processes/{processId}:
    get:
      tags: [public]
      operationId: publicNodechainByProcess
      security: []
      summary: Process-scoped journal history
      parameters:
        - $ref: '#/components/parameters/ProcessId'
        - name: limit
          in: query
          schema:
            type: integer
            maximum: 500
      responses:
        '200':
          description: Records list
...
  /v1/auth/login:
    post:
      tags: [auth]
      operationId: login
      security: []
      summary: Institution login (allowlisted token) → session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Session issued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /v1/auth/me:
    get:
      tags: [auth]
      operationId: me
      summary: Current session
      responses:
        '200':
          description: Session info
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionInfo'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /v1/processes:
    get:
      tags: [processes]
      operationId: listProcesses
      summary: List edge-tracked processes for session institution
      parameters:
        - $ref: '#/components/parameters/SessionId'
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/EdgeProcessStatus'
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
      responses:
        '200':
          description: Process list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProcessList'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      tags: [processes]
      operationId: createProcess
      summary: Submit process (legacy path; same as tokenization/start)
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/SessionId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProcessRequest'
      responses:
        '202':
          description: Accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProcessResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/Unprocessable'
  /v1/processes/stats:
    get:
      tags: [processes]
      operationId: processStats
      summary: Dashboard KPIs for institution
      parameters:
        - $ref: '#/components/parameters/SessionId'
      responses:
        '200':
          description: Stats
        '401':
          $ref: '#/components/responses/Unauthorized'
  /v1/processes/{processId}:
    get:
      tags: [processes]
      operationId: getProcess
      summary: Process status (edge + Core merge when available)
      parameters:
        - $ref: '#/components/parameters/ProcessId'
        - $ref: '#/components/parameters/SessionId'
      responses:
        '200':
          description: Current process state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProcessResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /v1/documents/hash:
    post:
      tags: [documents]
      operationId: hashDocuments
      summary: SHA-256 document package at the edge
      parameters:
        - $ref: '#/components/parameters/SessionId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HashDocumentsRequest'
      responses:
        '200':
          description: Hash computed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HashDocumentsResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/Unprocessable'
  /v1/documents/upload:
    post:
      tags: [documents]
      operationId: uploadDocument
      summary: Upload document evidence with signature attestation
      description: |
        Accepts file content (or precomputed hash) plus signature material.
        v1 verifies non-empty signature attestation + computes SHA-256 package hash.
        Full X.509 КЭП chain validation is follow-on.
      parameters:
        - $ref: '#/components/parameters/SessionId'
        - $ref: '#/components/parameters/IdempotencyKeyOptional'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DocumentUploadRequest'
          multipart/form-data:
            schema:
              type: object
              required: [processId, signature]
              properties:
                file:
                  type: string
                  format: binary
                processId:
                  type: string
                  description: AST processId
                signature:
                  type: string
                  description: Base64-encoded signature material (КЭП attestation)
                fileName:
                  type: string
      responses:
        '200':
          description: Document package recorded on edge
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentUploadResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/Unprocessable'
...
components:
  securitySchemes:
    SessionAuth:
      type: apiKey
      in: header
      name: X-Session-Id
      description: Session from POST /v1/auth/login (v1 demo / local)
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Target production institution OIDC/JWT
    MutualTLS:
      type: mutualTLS
      description: Target production institution mTLS
  parameters:
    SessionId:
      name: X-Session-Id
      in: header
      required: true
      schema:
        type: string
        minLength: 16
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        minLength: 8
        maxLength: 128
    IdempotencyKeyOptional:
      name: Idempotency-Key
      in: header
      required: false
      schema:
        type: string
        minLength: 8
        maxLength: 128
    ProcessId:
      name: processId
      in: path
      required: true
      schema:
        type: string
        pattern: '^AST-[A-Za-z0-9]+-\d{8}-[A-Za-z0-9]+$'
        example: AST-DEMO-20260720-a1b2c3d4e5f6
  schemas:
...
    LoginRequest:
      type: object
      required: [institutionId, token]
      properties:
        institutionId:
          type: string
          example: DEMO
        token:
          type: string
    LoginResponse:
      type: object
      required: [sessionId, institutionId, displayName, expiresAt]
      properties:
        sessionId:
          type: string
        institutionId:
          type: string
        displayName:
          type: string
        expiresAt:
          type: string
          format: date-time
    SessionInfo:
      type: object
      required: [institutionId, displayName, expiresAt]
      properties:
        institutionId:
          type: string
        displayName:
          type: string
        expiresAt:
          type: string
          format: date-time
        product:
          type: string
        capabilities:
          type: object
          properties:
            primaryTokenization:
              type: boolean
            documentHash:
              type: boolean
            coreHandOff:
              type: boolean
            mintOnEdge:
              type: boolean
              enum: [false]
...
    EdgeProcessStatus:
      type: string
      enum:
        - accepted
        - awaiting_core
        - documents_pending
        - submitted_to_core
        - rejected
        - duplicate
        - created
        - pot_pending
        - completed
        - failed
        - expired
    AssetType:
      type: string
      enum: [real_estate, bond, investment_package, other]
...
    CreateProcessRequest:
      type: object
      required:
        - processType
        - valuation
        - holderId
        - hasQualifiedSignature
        - documentPackageHash
      properties:
        processType:
          type: string
          enum: [primary_tokenization, revaluation, ownership_transfer]
        valuation:
          type: string
          pattern: '^-?\d+(\.\d{1,9})?$'
        holderId:
          type: string
        assetId:
          type: string
        hasQualifiedSignature:
          type: boolean
        documentPackageHash:
          type: string
          pattern: '^[a-fA-F0-9]{64}$'
        processId:
          type: string
        note:
          type: string
    ProcessResponse:
      type: object
      properties:
        processId:
          type: string
          example: AST-DEMO-20260720-a1b2c3d4e5f6
        claimId:
          type: string
          nullable: true
          description: Claim id (v1 equals processId)
        status:
          $ref: '#/components/schemas/EdgeProcessStatus'
        currentStep:
          type: string
          description: Human/pipeline step label
        institutionId:
          type: string
        valuation:
          type: string
        holderId:
          type: string
        assetType:
          $ref: '#/components/schemas/AssetType'
        documentPackageHash:
          type: string
        idempotencyKey:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        source:
          type: string
          enum: [edge, core]
        message:
          type: string
        core:
          type: object
          additionalProperties: true
        coreError:
          type: object
          additionalProperties: true
    ProcessList:
      type: object
      required: [institutionId, count, processes]
      properties:
        institutionId:
          type: string
        count:
          type: integer
        processes:
          type: array
          items:
            $ref: '#/components/schemas/ProcessResponse'
    HashDocumentsRequest:
      type: object
      properties:
        rawPackage:
          type: string
        parts:
          type: array
          items:
            type: object
            required: [name, content]
            properties:
              name:
                type: string
              content:
                type: string
              encoding:
                type: string
                enum: [utf8, base64]
    HashDocumentsResponse:
      type: object
      required: [documentPackageHash, institutionId]
      properties:
        documentPackageHash:
          type: string
        institutionId:
          type: string
        byteLength:
          type: integer
        partCount:
          type: integer
    DocumentUploadRequest:
      type: object
      required: [processId, signature]
      properties:
        processId:
          type: string
        signature:
          type: string
          description: Base64 signature material or attestation token
        fileName:
          type: string
        contentBase64:
          type: string
          description: Optional file bytes as base64; used to compute package hash
        documentPackageHash:
          type: string
          description: Optional precomputed SHA-256 hex (64 chars)
        hasQualifiedSignature:
          type: boolean
          default: true
    DocumentUploadResponse:
      type: object
      properties:
        documentId:
          type: string
        processId:
          type: string
        documentPackageHash:
          type: string
        status:
          type: string
          enum: [accepted, rejected]
        rejectionReason:
          type: string
          nullable: true
...
    ErrorBody:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: array
          items:
            type: string
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorBody'
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorBody'
...
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorBody'
...
    Unprocessable:
      description: Validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorBody'